requirements.txt
# ── PMC Overwatch — Pinned Dependencies ────────────────────────────
# Generated from a clean Python 3.13 venv on macOS ARM64.
# Only direct dependencies are listed; transitive deps are managed by pip.

opencv-python==4.13.0.92
twitchio==2.10.0
SpeechRecognition==3.14.5
sounddevice==0.5.5
soundfile==0.13.1
numpy==2.4.2
ollama==0.6.1
python-dotenv==1.2.1
customtkinter==5.2.2
